feat(core/unpublish): db revisions not updated
Only the first batch size revisions are updated in the database. The reason: inside the transaction doctrine clear is called, next batch we pass a detached entity to the bulkUnpublish method. This method removes the environment with a detached entity.

Before starting a new transaction, fetch the environment object again.