Attempting to change readonly property EMS\SubmissionBundle\Entity\FormSubmission::$id
Because we create the id or self in the constructor for each submission. We should not use a generator strategy.

The support for read only properties only works fine for entities which use the "None" strategy for identifier generator. When the entity use other strategy, like the suggested in the documentation, an exception is thrown when you try to remove the entity.