fix(admin/doctrine): invalid entity mapping for new EnvironmentRevision
The 'inversed-by' were missing leading to the following errors:

The field EMS\CoreBundle\Entity\Environment#environmentRevisions is on the inverse side of a bi-directional relationship, but the specified mappedBy association on the target-entity EMS\CoreBundle\Entity\EnvironmentRevision#environment does not contain the required 'inversedBy="environmentRevisions"' attribute.

The field EMS\CoreBundle\Entity\Revision#environmentRevisions is on the inverse side of a bi-directional relationship, but the specified mappedBy association on the target-entity EMS\CoreBundle\Entity\EnvironmentRevision#revision does not contain the required 'inversedBy="environmentRevisions"' attribute.